+++ app/src/main/java/es/jvbabi/vplanplus/domain/repository/TeacherRepository.kt
@@ -5,12 +5,12 @@ import es.jvbabi.vplanplus.domain.model.Teacher
 import java.util.UUID
 
 interface TeacherRepository {
-    suspend fun createTeacher(schoolId: Long, acronym: String)
-    suspend fun getTeachersBySchoolId(schoolId: Long): List<Teacher>
+    suspend fun createTeacher(schoolId: Int, acronym: String)
+    suspend fun getTeachersBySchoolId(schoolId: Int): List<Teacher>
     suspend fun find(school: School, acronym: String, createIfNotExists: Boolean = false): Teacher?
     suspend fun getTeacherById(id: UUID): Teacher?
-    suspend fun deleteTeachersBySchoolId(schoolId: Long)
-    suspend fun insertTeachersByAcronym(schoolId: Long, teachers: List<String>)
+    suspend fun deleteTeachersBySchoolId(schoolId: Int)
+    suspend fun insertTeachersByAcronym(schoolId: Int, teachers: List<String>)
 
     suspend fun getAll(): List<Teacher>
 }
